+++ components/select/index.jsx
@@ -1,36 +1,37 @@
 import React from 'react';
 import Select from 'rc-select';
+import classNames from 'classnames';
 
-let AntSelect = React.createClass({
+const AntSelect = React.createClass({
   getDefaultProps() {
     return {
       prefixCls: 'ant-select',
       transitionName: 'slide-up',
       optionLabelProp: 'children',
+      choiceTransitionName: 'zoom',
       showSearch: false,
       size: 'default'
     };
   },
   render() {
-    const {size, className, combobox, notFoundContent} = this.props;
+    let {
+      size, className, combobox, notFoundContent
+    } = this.props;
 
-    let sizeClass = null;
-    if (size === 'large') {
-      sizeClass = 'ant-select-lg';
-    } else if (size === 'small') {
-      sizeClass = 'ant-select-sm';
+    const cls = classNames({
+      'ant-select-lg': size === 'large',
+      'ant-select-sm': size === 'small',
+      [className]: !!className,
+    });
+
+    if (combobox) {
+      notFoundContent = null;
     }
 
-    const classNames = [];
-
-    if (className) {
-      classNames.push(className);
-    }
-    if (sizeClass) {
-      classNames.push(sizeClass);
-    }
     return (
-      <Select {...this.props} className={classNames.join(' ')} notFoundContent={combobox ? null : notFoundContent} />
+      <Select {...this.props}
+        className={cls}
+        notFoundContent={notFoundContent} />
     );
   }
 });