+++ components/table/index.jsx
@@ -1,5 +1,4 @@
 import React from 'react';
-import reqwest from 'reqwest';
 import Table from 'rc-table';
 import Checkbox from '../checkbox';
 import Radio from '../radio';
@@ -8,47 +7,24 @@ import Pagination from '../pagination';
 import Icon from '../icon';
 import objectAssign from 'object-assign';
 import Spin from '../spin';
+import classNames from 'classnames';
 
 function noop() {
 }
 
-function defaultResolve(data) {
-  return data || [];
-}
-
-class DataSource {
-  init(config) {
-    this.config = config;
-    this.url = config.url || '';
-    this.resolve = config.resolve || defaultResolve;
-    this.getParams = config.getParams || noop;
-    this.getPagination = config.getPagination || noop;
-    this.headers = config.headers || {};
-    this.data = config.data || {};
-  }
-
-  constructor(config) {
-    if (config) {
-      this.init(config);
-    }
-  }
-
-  clone(config = {}) {
-    return new DataSource(objectAssign({}, this.config, config));
-  }
-}
+const defaultLocale = {
+  filterTitle: '筛选',
+  filterConfirm: '确定',
+  filterReset: '重置'
+};
 
 let AntTable = React.createClass({
   getInitialState() {
     return {
       // 减少状态
       selectedRowKeys: [],
-      // only for remote
-      data: [],
-      dataSource: this.props.dataSource,
       filters: {},
       selectionDirty: false,
-      loading: this.props.loading,
       sortColumn: '',
       sortOrder: '',
       sorter: null,
@@ -62,20 +38,30 @@ let AntTable = React.createClass({
 
   getDefaultProps() {
     return {
+      dataSource: [],
       prefixCls: 'ant-table',
       useFixedHeader: false,
       rowSelection: null,
       className: '',
-      size: 'default',
+      size: 'large',
       loading: false,
       bordered: false,
-      onChange: function () {
-      }
+      onChange: noop,
+      locale: {}
     };
   },
 
   propTypes: {
-    dataSource: React.PropTypes.oneOfType([React.PropTypes.array, React.PropTypes.instanceOf(DataSource)])
+    dataSource: React.PropTypes.array,
+    prefixCls: React.PropTypes.string,
+    useFixedHeader: React.PropTypes.bool,
+    rowSelection: React.PropTypes.object,
+    className: React.PropTypes.string,
+    size: React.PropTypes.string,
+    loading: React.PropTypes.bool,
+    bordered: React.PropTypes.bool,
+    onChange: React.PropTypes.func,
+    locale: React.PropTypes.object,
   },
 
   getDefaultSelection() {
@@ -105,43 +91,22 @@ let AntTable = React.createClass({
         nextProps.dataSource !== this.props.dataSource) {
       let selectedRowKeys = this.state.selectedRowKeys;
       // 把不在当前页的选中项去掉
-      if (this.isLocalDataSource()) {
-        let currentPageRowKeys = this.getLocalDataPaging(nextProps.dataSource);
-        selectedRowKeys = selectedRowKeys.filter((key) => {
-          return currentPageRowKeys.indexOf(key) >= 0;
-        });
-      }
+      let currentPageRowKeys =
+        this.getCurrentPageData(nextProps.dataSource).map(
+          (record, i) => this.getRecordKey(record, i)
+        );
+      selectedRowKeys = selectedRowKeys.filter((key) => {
+        return currentPageRowKeys.indexOf(key) >= 0;
+      });
       this.setState({
         selectionDirty: false,
-        dataSource: nextProps.dataSource,
-        loading: true
-      }, this.fetch);
-    }
-    if (nextProps.columns !== this.props.columns) {
-      this.setState({
-        filters: {}
-      });
-    }
-    if ('loading' in nextProps) {
-      this.setState({
-        loading: nextProps.loading
+        selectedRowKeys,
       });
     }
   },
 
   hasPagination(pagination) {
-    if (pagination === undefined) {
-      pagination = this.props.pagination;
-    }
-    return pagination !== false;
-  },
-
-  isLocalDataSource() {
-    return Array.isArray(this.state.dataSource);
-  },
-
-  getRemoteDataSource() {
-    return this.state.dataSource;
+    return this.props.pagination !== false;
   },
 
   toggleSortOrder(order, column) {
@@ -161,7 +126,7 @@ let AntTable = React.createClass({
         sortOrder = order;
       }
     }
-    if (this.isLocalDataSource()) {
+    if (typeof column.sorter === 'function') {
       sorter = function () {
         let result = column.sorter.apply(this, arguments);
         if (sortOrder === 'ascend') {
@@ -176,7 +141,7 @@ let AntTable = React.createClass({
       sortColumn,
       sorter
     };
-    this.fetch(newState);
+    this.setState(newState);
     this.props.onChange.apply(this, this.prepareParamsArguments(objectAssign({}, this.state, newState)));
   },
 
@@ -184,12 +149,19 @@ let AntTable = React.createClass({
     filters = objectAssign({}, this.state.filters, {
       [this.getColumnKey(column)]: filters
     });
+    // Remove filters not in current columns
+    const currentColumnKeys = this.props.columns.map(c => this.getColumnKey(c));
+    Object.keys(filters).forEach((columnKey) => {
+      if (currentColumnKeys.indexOf(columnKey) < 0) {
+        delete filters[columnKey];
+      }
+    });
     const newState = {
       selectedRowKeys: [],
       selectionDirty: false,
       filters
     };
-    this.fetch(newState);
+    this.setState(newState);
     this.props.onChange.apply(this, this.prepareParamsArguments(objectAssign({}, this.state, newState)));
   },
 
@@ -280,7 +252,7 @@ let AntTable = React.createClass({
       selectionDirty: false,
       pagination
     };
-    this.fetch(newState);
+    this.setState(newState);
     this.props.onChange.apply(this, this.prepareParamsArguments(objectAssign({}, this.state, newState)));
   },
 
@@ -356,7 +328,6 @@ let AntTable = React.createClass({
       if (this.props.rowSelection.type === 'radio') {
         selectionColumn = {
           key: 'selection-column',
-          width: 60,
           render: this.renderSelectionRadio,
           className: 'ant-table-selection-column'
         };
@@ -365,7 +336,6 @@ let AntTable = React.createClass({
         selectionColumn = {
           key: 'selection-column',
           title: checkboxAll,
-          width: 60,
           render: this.renderSelectionCheckBox,
           className: 'ant-table-selection-column'
         };
@@ -378,10 +348,6 @@ let AntTable = React.createClass({
       }
     }
     return columns;
-  },
-
-  getCurrentPageData() {
-    return this.isLocalDataSource() ? this.getLocalDataPaging() : this.state.data;
   },
 
   getColumnKey(column, index) {
@@ -398,6 +364,7 @@ let AntTable = React.createClass({
   },
 
   renderColumnsDropdown(columns) {
+    let locale = objectAssign({}, defaultLocale, this.props.locale);
     return columns.map((column, i) => {
       column = objectAssign({}, column);
       let key = this.getColumnKey(column, i);
@@ -405,7 +372,7 @@ let AntTable = React.createClass({
       if (column.filters && column.filters.length > 0) {
         let colFilters = this.state.filters[key] || [];
         filterDropdown =
-          <FilterDropdown column={column}
+          <FilterDropdown locale={locale} column={column}
                           selectedKeys={colFilters}
                           confirmFilter={this.handleFilter}/>;
       }
@@ -417,16 +384,17 @@ let AntTable = React.createClass({
             column.className += ' ant-table-column-sort';
           }
         }
+
         sortButton = <div className="ant-table-column-sorter">
           <span className={'ant-table-column-sorter-up ' +
                            ((isSortColumn && this.state.sortOrder === 'ascend') ? 'on' : 'off')}
-                title="升序排序"
+                title="↑"
                 onClick={this.toggleSortOrder.bind(this, 'ascend', column)}>
             <Icon type="caret-up"/>
           </span>
           <span className={'ant-table-column-sorter-down ' +
                            ((isSortColumn && this.state.sortOrder === 'descend') ? 'on' : 'off')}
-                title="降序排序"
+                title="↓"
                 onClick={this.toggleSortOrder.bind(this, 'descend', column)}>
             <Icon type="caret-down"/>
           </span>
@@ -445,7 +413,7 @@ let AntTable = React.createClass({
     let pagination = objectAssign(this.state.pagination, {
       pageSize: pageSize
     });
-    this.fetch({pagination});
+    this.setState({ pagination });
   },
 
   renderPagination() {
@@ -453,34 +421,25 @@ let AntTable = React.createClass({
     if (!this.hasPagination()) {
       return null;
     }
-    let classString = 'ant-table-pagination';
-    if (this.props.size === 'small') {
-      classString += ' mini';
-    }
-    let total = this.state.pagination.total;
-    if (!total && this.isLocalDataSource()) {
-      total = this.getLocalData().length;
-    }
-    return (total > 0) ? <Pagination className={classString}
-                                     onChange={this.handlePageChange}
-                                     total={total}
-                                     pageSize={10}
-                                     onShowSizeChange={this.handleShowSizeChange}
-      {...this.state.pagination} /> : null;
+    let classString = classNames({
+      'ant-table-pagination': true,
+      'mini': this.props.size === 'middle' || this.props.size === 'small',
+    });
+    let total = this.state.pagination.total || this.getLocalData().length;
+    return (total > 0) ?
+      <Pagination className={classString}
+                  onChange={this.handlePageChange}
+                  total={total}
+                  pageSize={10}
+                  onShowSizeChange={this.handleShowSizeChange}
+                  {...this.state.pagination} /> : null;
   },
 
   prepareParamsArguments(state) {
     // 准备筛选、排序、分页的参数
-    let pagination;
-    let filters = {};
-    let sorter = {};
-    pagination = state.pagination;
-    this.props.columns.forEach((column) => {
-      let colFilters = state.filters[this.getColumnKey(column)] || [];
-      if (colFilters.length > 0) {
-        filters[this.getColumnKey(column)] = colFilters;
-      }
-    });
+    const pagination = state.pagination;
+    const filters = state.filters;
+    const sorter = {};
     if (state.sortColumn &&
       state.sortOrder &&
       state.sortColumn.dataIndex) {
@@ -490,61 +449,13 @@ let AntTable = React.createClass({
     return [pagination, filters, sorter];
   },
 
-  fetch(newState) {
-    if (this.isLocalDataSource()) {
-      if (newState) {
-        this.setState(newState);
-      }
-    } else {
-      // remote 模式使用 this.dataSource
-      let dataSource = this.getRemoteDataSource();
-      if (!dataSource) {
-        return null;
-      }
-      let state = objectAssign({}, this.state, newState);
-      if (newState || !this.state.loading) {
-        this.setState(objectAssign({
-          loading: true
-        }, newState));
-      }
-      let buildInParams = dataSource.getParams.apply(this, this.prepareParamsArguments(state)) || {};
-      return reqwest({
-        url: dataSource.url,
-        method: 'get',
-        data: objectAssign(buildInParams, dataSource.data),
-        headers: dataSource.headers,
-        type: 'json',
-        success: (result) => {
-          if (this.isMounted()) {
-            let pagination = objectAssign(
-              state.pagination,
-              dataSource.getPagination.call(this, result)
-            );
-            this.setState({
-              selectionDirty: false,
-              loading: false,
-              data: dataSource.resolve.call(this, result),
-              pagination: pagination
-            });
-          }
-        },
-        error: () => {
-          this.setState({
-            loading: false,
-            data: []
-          });
-        }
-      });
-    }
-  },
-
   findColumn(myKey) {
     return this.props.columns.filter((c) => {
       return this.getColumnKey(c) === myKey;
     })[0];
   },
 
-  getLocalDataPaging(dataSource) {
+  getCurrentPageData(dataSource) {
     let data = this.getLocalData(dataSource);
     let current, pageSize;
     let state = this.state;
@@ -573,7 +484,7 @@ let AntTable = React.createClass({
 
   getLocalData(dataSource) {
     let state = this.state;
-    let data = dataSource || this.state.dataSource;
+    let data = dataSource || this.props.dataSource;
     // 排序
     if (state.sortOrder && state.sorter) {
       data = data.sort(state.sorter);
@@ -582,40 +493,35 @@ let AntTable = React.createClass({
     if (state.filters) {
       Object.keys(state.filters).forEach((columnKey) => {
         let col = this.findColumn(columnKey);
+        if (!col) {
+          return;
+        }
         let values = state.filters[columnKey] || [];
         if (values.length === 0) {
           return;
         }
-        data = data.filter((record) => {
-          return values.some((v)=> {
-            return col.onFilter(v, record);
-          });
-        });
+        data = col.onFilter ? data.filter(record => {
+          return values.some(v => col.onFilter(v, record));
+        }) : data;
       });
     }
     return data;
   },
 
-  componentDidMount() {
-    if (!this.isLocalDataSource()) {
-      this.fetch();
-    }
-  },
-
   render() {
     let data = this.getCurrentPageData();
     let columns = this.renderRowSelection();
-    let classString = this.props.className;
     let expandIconAsCell = this.props.expandedRowRender && this.props.expandIconAsCell !== false;
-    if (this.props.size === 'small') {
-      classString += ' ant-table-small';
-    }
-    if (this.props.bordered) {
-      classString += ' ant-table-bordered';
-    }
+
+    let classString = classNames({
+      [`ant-table-${this.props.size}`]: true,
+      'ant-table-bordered': this.props.bordered,
+      [this.props.className]: !!this.props.className,
+    });
+
     columns = this.renderColumnsDropdown(columns);
     columns = columns.map((column, i) => {
-      column.key = column.dataIndex || i;
+      column.key = column.key || column.dataIndex || i;
       return column;
     });
     let emptyText;
@@ -635,7 +541,7 @@ let AntTable = React.createClass({
         expandIconAsCell={expandIconAsCell} />
       {emptyText}
     </div>;
-    if (this.state.loading) {
+    if (this.props.loading) {
       // if there is no pagination or no data, the height of spin should decrease by half of pagination
       let paginationPatchClass = (this.hasPagination() && data && data.length !== 0)
               ? 'ant-table-with-pagination'
@@ -651,7 +557,5 @@ let AntTable = React.createClass({
     );
   }
 });
-
-AntTable.DataSource = DataSource;
 
 export default AntTable;