+++ components/upload/uploadList.jsx
@@ -3,10 +3,12 @@ import Animate from 'rc-animate';
 import Icon from '../icon';
 const prefixCls = 'ant-upload';
 import { Line } from '../progress';
+import classNames from 'classnames';
 
 export default React.createClass({
   getDefaultProps() {
     return {
+      listType: 'text',  // or picture
       items: [],
       progressAttr: {
         strokeWidth: 3,
@@ -20,26 +22,40 @@ export default React.createClass({
   render() {
     let list = this.props.items.map((file) => {
       let progress;
-      let infoUploadingClass = '';
+      let icon = <Icon type="paper-clip" />;
+
+      if (this.props.listType === 'picture') {
+        icon = (file.status === 'uploading' || !file.url)
+          ? <Icon className={prefixCls + '-list-item-thumbnail'} type="picture" />
+          : <a className={prefixCls + '-list-item-thumbnail'}
+               href={file.url}
+               target="_blank"><img src={file.url} alt={file.name} /></a>;
+      }
       if (file.status === 'uploading') {
         progress = <div className={prefixCls + '-list-item-progress'}>
           <Line {...this.props.progressAttr} percent={file.percent} />
         </div>;
-        infoUploadingClass = ' ' + prefixCls + '-list-item-uploading';
       }
+      const infoUploadingClass = classNames({
+        [`${prefixCls}-list-item`]: true,
+        [`${prefixCls}-list-item-${file.status}`]: true,
+      });
       return (
-        <div className={prefixCls + '-list-item' + infoUploadingClass} key={file.uid}>
+        <div className={infoUploadingClass} key={file.uid}>
           <div className={prefixCls + '-list-item-info'}>
-            <Icon type="paper-clip" />
-            <span className="ant-upload-item-name">{file.name}</span>
-            <Icon type="cross" ref="theCloseBtn"
-              onClick={this.handleClose.bind(this, file)} />
+            {icon}
+            <span className={prefixCls + '-list-item-name'}>{file.name}</span>
+            <Icon type="cross" onClick={this.handleClose.bind(this, file)} />
           </div>
           { progress }
         </div>
       );
     });
-    return <div className={prefixCls + '-list'}>
+    const listClassNames = classNames({
+      [`${prefixCls}-list`]: true,
+      [`${prefixCls}-list-${this.props.listType}`]: true,
+    });
+    return <div className={listClassNames}>
       <Animate transitionName={prefixCls + '-margin-top'}>
         {list}
       </Animate>